Add retrieveCommands utility
